requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
boto3==1.16.16
botocore==1.19.16
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
discord==1.0.1
discord.py==1.7.1
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
multidict==5.1.0
numpy==1.19.4
pefile==2019.4.18
Pillow==8.0.1
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.10
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
stripe==2.55.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.6.3
